Big changes
After much deliberation, I have decided to not take a dependency on RestSharp. The following are the reasons:

1) RestSharp is no longer one 100 KB dll that can pulled into the project like WP8
2) They do not have a WP8.1 flavor and they don't have a UWP PCL one yet
3) People are relying on RestSharp.Portable which is a fork from the RestSharp codebase that has been made to work for UWP apps.
4) Including RestSharp.Portable pulls in so many other dependencies like Bcl Async, BouncyCastle Crypto,  HttpUtility, NewtonSoft JSON etc.
5) These are all big over kills for this app and can "potentially" slow down the start up time.

There is lot of emphasis on performance so I'm not doing this. Instead I decided to use native c# HttpClient and just provide an interface for authenticating requests. If we don't authenticate this will just do a standard web request. But if we do authenticate, we will add some auth headers to the request. The full sign in flow has been verified and is working fine.

Another advantage with this approach is a lot more code can be pushed down to the model layer. Hence lesser to implement on the View.Shared side.

Renamed Epiphany.Model.Web to just Epiphany.Web

Introduced test projects for every platform we compile. But unfortunately, I', not able to get CompareNetObjects to work on the test projects. So these have been put on hold for the time being.