Design the offline experience.
We will show some text and a log in button for the Home Tab and for the My Books Tab. We will a background image as there is not much content on these pages and it looks bare. Hook up the Log in button to the LogonPage and LogonPage will just close itself after completing now.