Improve performance of DataSource.
DataSource today executes the web request and gets back the response object. Then it reflects on the object to check if there is an object of the type we expect in the response that is non-null. This is too slow.

DataSource will not expose a Func which consumers of DataSource will set to get the correct object from the Response object.