Add OrderBy ability to books in a shelf.
Add converters for the proper text for each sort type and sort order and add appropriate strings. The strings for order by will change based upon the sort paradigm. So the converter that handles that now knows what the current sort by filter is. Based upon that, strings will be provided e.g. Highest or Lowest for numbers, Ascending or Descending for strings. Newest or Oldest for dates etc.